LayoutTests/fast/css/pseudo-active-on-labeled-control-without-renderer-expected.txt

Verify that a labeled elemed gets the :active state even if it has no renderer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



    Initial state
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
On Mouse Down
PASS elementsWithActiveStyle() is ["target", "labelable"]
PASS elementsMatchingActiveSelector() is ["html", "body", "webkit-test", "labelable-ancestor", "labelable-parent", "target", "labelable"]
After Mouse Down
PASS elementsWithActiveStyle() is ["target", "labelable"]
PASS elementsMatchingActiveSelector() is ["html", "body", "webkit-test", "labelable-ancestor", "labelable-parent", "target", "labelable"]
On Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
After Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
On Mouse Down
PASS elementsWithActiveStyle() is ["target", "labelable"]
PASS elementsMatchingActiveSelector() is ["html", "body", "webkit-test", "labelable-ancestor", "labelable-parent", "target", "labelable"]
After Mouse Down
PASS elementsWithActiveStyle() is ["target", "labelable"]
PASS elementsMatchingActiveSelector() is ["html", "body", "webkit-test", "labelable-ancestor", "labelable-parent", "target", "labelable"]
On Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
After Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
On Mouse Down
PASS elementsWithActiveStyle() is ["target", "labelable"]
PASS elementsMatchingActiveSelector() is ["html", "body", "webkit-test", "labelable-ancestor", "labelable-parent", "target", "labelable"]
After Mouse Down
PASS elementsWithActiveStyle() is ["target", "labelable"]
PASS elementsMatchingActiveSelector() is ["html", "body", "webkit-test", "labelable-ancestor", "labelable-parent", "target", "labelable"]
On Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
After Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
PASS successfullyParsed is true

TEST COMPLETE
